requirements.txt
# ------------ Django ----------------
django 
djangorestframework # 用于构建RESTful API
Pillow 
django-cors-headers # 用于解决跨域问题
drf-yasg # 用于构建Swagger文档
# ------------------------------------

# ---------------- ai ----------------
openai
# ------------------------------------

# -------------- 数据库 ---------------
mysqlclient
PyMySQL
# ------------------------------------

# -------------  代码风格检查工具 -----------
flake8
flake8-django
flake8-docstrings
# ----------------------------------------

# ------------ 文字相关搜索 ----------------
jieba
scikit-learn
# ----------------------------------------

